# nossa-igreja-specification/src/main/resources/swagger.yaml
swagger: "2.0"
info:
  description: "API de gerenciamento de doações e grupos de comunhão para igrejas"
  version: "1.0.0"
  title: "Nossa igreja - REST API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "andressa.obm@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/api/v1"
...
schemes:
- "https"
- "http"
paths:
  /users:
    post:
      tags:
      - "user"
      summary: "Cadastra um usuário"
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "userCadastroDTO"
        required: true
        schema:
          $ref: "#/definitions/UserCadastroDTO"
      responses:
        "201":
          description: "created"
          schema:
            $ref: "#/definitions/UserDTO"
        "400":
          description: "bad request"
    put:
      tags:
      - "user"
      summary: "Atualiza dados do usuário"
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "userDTO"
        required: true
        schema:
          $ref: "#/definitions/UserDTO"
      responses:
        "400":
          description: "bad request"
        "404":
          description: "not found"
      security:
        - auth:
          - "oauth2"
  /users/{id}:
    get:
      tags:
      - "user"
      summary: "Usuário por id"
      description: ""
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/UserDTO"
        "400":
          description: "bad request"
        "404":
          description: "not found"
      security:
        - auth:
          - "oauth2"
    delete:
      tags:
      - "user"
      summary: "Exclui um usuário"
      description: ""
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
        "400":
          description: "bad request"
        "404":
          description: "not found"
      security:
        - auth:
          - "oauth2"
...
securityDefinitions:
  auth:
    type: "oauth2"
    authorizationUrl: "http://localhost8082/auth/"
    flow: "implicit"
    scopes:
      oauth2: "authorization"
definitions:
...
  UserDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      firstName:
        type: "string"
      lastName:
        type: "string"
      birthDate:
        type: "string"
        format: "date"
      phoneNumber:
        type: "string"
      email:
        type: "string"
  UserCadastroDTO:
    type: "object"
    properties:
      firstName:
        type: "string"
      lastName:
        type: "string"
      birthDate:
        type: "string"
        format: "date"
      phoneNumber:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      acceptTerms:
        type: "boolean"
...